# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "E-Pharmacy app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is the documentation of E-Pharmacy - a web platform for online pharmacies."
  },
  "servers": [
    {
      "url": null
    },
    {
      "url": "http://localhost:4000/api"
    }
  ],
...
  "paths": {
...
    "/api/stores/nearest": {
      "get": {
        "tags": [
          "Pharmacies"
        ],
        "summary": "Get nearest pharmacies",
        "operationId": "getPharmaciesNearest",
        "description": "Public endpoint. Returns pharmacies near given coordinates. Supports pagination.",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number",
              "example": 49.5535
            },
            "description": "Latitude"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number",
              "example": 25.5948
            },
            "description": "Longitude"
          },
          {
            "in": "query",
            "name": "radius",
            "required": false,
            "schema": {
              "type": "number",
              "default": 2000,
              "example": 2000
            },
            "description": "Search radius in meters"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 12
            },
            "description": "Number of pharmacies per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Nearest pharmacies fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data",
                    "page",
                    "limit",
                    "totalItems",
                    "totalPages",
                    "hasNextPage",
                    "hasPreviousPage"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Nearest pharmacies fetched successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/pharmacy"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "distance": {
                                "type": "number",
                                "description": "Distance to pharmacy in meters",
                                "example": 350.5
                              }
                            }
                          }
                        ]
                      }
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 12
                    },
                    "totalItems": {
                      "type": "integer",
                      "example": 5
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 1
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "example": false
                    },
                    "hasPreviousPage": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "pharmacy": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "address",
          "city",
          "phone",
          "rating",
          "location"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Pharmacy ID",
            "example": "68fad270e7f1469dd713faef"
          },
          "name": {
            "type": "string",
            "description": "Pharmacy name",
            "example": "Balzam"
          },
          "address": {
            "type": "string",
            "description": "Pharmacy address",
            "example": "Shevchenka 12"
          },
          "city": {
            "type": "string",
            "description": "Pharmacy city",
            "example": "Ternopil"
          },
          "phone": {
            "type": "string",
            "description": "Pharmacy phone",
            "example": "+380994056328"
          },
          "rating": {
            "type": "number",
            "description": "Pharmacy rating",
            "example": 4
          },
          "location": {
            "type": "object",
            "required": [
              "type",
              "coordinates"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Point"
                ],
                "description": "Location pharmacy",
                "example": "Point"
              },
              "coordinates": {
                "type": "array",
                "items": {
                  "type": "array",
                  "description": "Coordinates pharmacy",
                  "example": [
                    25.5948,
                    49.5535
                  ]
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Document creation date",
            "example": "2026-04-07T12:34:56.789Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date",
            "example": "2026-04-07T12:34:56.789Z"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequestError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message",
                    "errors"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Name should have at least 3 characters"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "example": "['name']"
                            }
                          },
                          "type": {
                            "type": "string",
                            "example": "string.min"
                          },
                          "context": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server error"
                }
              }
            }
          }
        }
      }
    }
  }
}